% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{W}{integer (maximum weight fitting in the knapsack)}

\item{parallel}{bool}
}
\value{
Returns a list object containing value and elements information.
}
\description{
Implementation of knapsack problem using a brute force approach.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
