% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame}

\item{W}{integer (maximum weight fitting in the knapsack)}
}
\value{
Returns a list object containing value and elements information.
}
\description{
Implementation of knapsack problem using greedy algorithm.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
