% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic Programming Knapsack}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame}

\item{W}{integer (maximum weight fitting in the knapsack)}
}
\value{
Returns a list object containing value and elements information.
}
\description{
Implementation of knapsack problem using dynamic programming.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
